requirements.txt
streamlit
requests
pandas
plotly
supabase
python-dotenv